requirements.txt
# Core dependencies for Financial RAG System
requests>=2.31.0
python-dotenv>=1.0.0

# Document processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2024.2.26

# ML and embeddings
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector storage
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Optional LLM integrations
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0

# Utilities
tqdm>=4.66.0
pandas>=2.0.0
python-dateutil>=2.8.0

# Development (optional)
jupyter>=1.0.0
ipython>=8.0.0